requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
eval_type_backport==0.2.2
fastapi==0.119.1
frozenlist==1.8.0
greenlet==3.2.4
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
invoke==2.2.1
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-groq==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
marshmallow==3.26.1
mistralai==1.9.11
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.5.0
opentelemetry-api==1.38.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pillow==12.0.0
playwright==1.55.0
propcache==0.4.1
protobuf==5.29.5
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
texttable==1.7.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
